+++ src/main/java/wooteco/subway/line/dao/LineDao.java
@@ -108,6 +108,6 @@ public class LineDao {
     }
 
     public void deleteById(Long id) {
-        jdbcTemplate.update("delete from Line where id = ?", id);
+        jdbcTemplate.update("delete from LINE where id = ?", id);
     }
 }
