+++ src/main/java/wooteco/subway/path/exception/FareCalculationException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.path.application;
+package wooteco.subway.path.exception;
 
 public class FareCalculationException extends RuntimeException {
     private static final String ERROR_MESSAGE = "요금 계산에 실패했습니다.";
