+++ src/main/java/wooteco/subway/path/exception/InvalidPathException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.path.application;
+package wooteco.subway.path.exception;
 
 public class InvalidPathException extends RuntimeException {
     public InvalidPathException(String errorMessage) {
