+++ src/main/java/wooteco/subway/path/exception/PathNotConnectedException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.path.application;
+package wooteco.subway.path.exception;
 
 public class PathNotConnectedException extends InvalidPathException {
     public PathNotConnectedException(String errorMessage) {
