+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -5,6 +5,8 @@ import wooteco.subway.station.dao.StationDao;
 import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationRequest;
 import wooteco.subway.station.dto.StationResponse;
+import wooteco.subway.station.exception.DuplicateStationNameException;
+import wooteco.subway.station.exception.StationNotFoundException;
 
 import java.util.List;
 import java.util.Optional;
