+++ src/main/java/wooteco/subway/station/exception/DuplicateStationNameException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.station.application;
+package wooteco.subway.station.exception;
 
 public class DuplicateStationNameException extends IllegalArgumentException {
     private static final String ERROR_MESSAGE = "이미 등록된 역입니다.";
