+++ src/main/java/wooteco/subway/station/exception/InvalidStationNameException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.station.domain;
+package wooteco.subway.station.exception;
 
 public class InvalidStationNameException extends IllegalArgumentException {
     public static final String ERROR_MESSAGE = "2자 이상 한글/숫자로 구성된 역 이름만 허용합니다.";
