+++ src/main/java/wooteco/subway/station/exception/StationNotFoundException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.station.application;
+package wooteco.subway.station.exception;
 
 public class StationNotFoundException extends IllegalArgumentException {
     private static final String ERROR_MESSAGE = "역이 등록되어 있지 않습니다.";
